Finish restructure: domain re-exports, split files, tidy discovery
- Populate every domain __init__.py with explicit feature re-exports + __all__
  so the cb.<domain>.Feature access path works (proposal §9), alongside the
  existing flat cb.feature.Feature path.
- Move InverseFourierTransform into its own transforms/inverse_fourier_transform.py
  (§5/§12); FFT helpers stay in fourier_transform.py. Regenerate feature.pyi.
- Split the combined wavelet test into test_feature_continuous_wavelet_transform.py
  and test_feature_discrete_wavelet_transform.py to match the source split.
- feature.py: wrap auto-discovery in a function so loop temporaries and helpers
  (PACKAGE_ROOT, Path) no longer leak onto the cb.feature namespace.
