+++ src/cobrabox/connectivity/__init__.py
@@ -1,1 +1,26 @@
-# Connectivity features: inter-channel relationships
+# Connectivity features: which regions are synchronized/interacting.
+from .coherence import Coherence
+from .correlation import Correlation
+from .covariance import Covariance
+from .directed_transfer_function import DirectedTransferFunction
+from .envelope_correlation import EnvelopeCorrelation
+from .granger_causality import GrangerCausality
+from .mutual_information import MutualInformation
+from .partial_correlation import PartialCorrelation
+from .partial_directed_coherence import PartialDirectedCoherence
+from .phase_locking_value import PhaseLockingValue
+from .reciprocal_connectivity import ReciprocalConnectivity
+
+__all__ = [
+    "Coherence",
+    "Correlation",
+    "Covariance",
+    "DirectedTransferFunction",
+    "EnvelopeCorrelation",
+    "GrangerCausality",
+    "MutualInformation",
+    "PartialCorrelation",
+    "PartialDirectedCoherence",
+    "PhaseLockingValue",
+    "ReciprocalConnectivity",
+]
