+++ src/cobrabox/decompositions/__init__.py
@@ -1,0 +1,5 @@
+# Decomposition features: break a signal into constituent components.
+from .emd import EMD
+from .svd import SVD
+
+__all__ = ["EMD", "SVD"]
