+++ src/cobrabox/feature.py
@@ -11,46 +11,55 @@ Two access paths:
 
 from __future__ import annotations
 
-import importlib
-from pathlib import Path
+import importlib as _importlib
+from pathlib import Path as _Path
 
-PACKAGE_ROOT = Path(__file__).parent
+_PACKAGE_ROOT = _Path(__file__).parent
 _PACKAGE_NAME = __package__ or "cobrabox"
 
 # Domains that should NOT be scanned for features.
 _DOMAIN_BLOCKLIST = {"egg", "__pycache__"}
 
-_discovered: dict[str, object] = {}
 
-for domain_dir in PACKAGE_ROOT.iterdir():
-    if not domain_dir.is_dir():
-        continue
-    if domain_dir.name in _DOMAIN_BLOCKLIST or domain_dir.name.startswith("_"):
-        continue
-    for module_path in domain_dir.rglob("*.py"):
-        if module_path.name == "__init__.py" or module_path.name.startswith("_"):
+def _discover() -> dict[str, object]:
+    """Import every domain module and collect classes flagged as features.
+
+    Wrapped in a function so loop temporaries do not leak into the module
+    namespace — only the discovered feature classes end up on ``cb.feature``.
+    """
+    discovered: dict[str, object] = {}
+    for domain_dir in _PACKAGE_ROOT.iterdir():
+        if not domain_dir.is_dir():
             continue
-        rel_path = module_path.relative_to(PACKAGE_ROOT)
-        module_name = ".".join(rel_path.with_suffix("").parts)
-        full_module_name = f"{_PACKAGE_NAME}.{module_name}"
-
-        try:
-            _module = importlib.import_module(full_module_name)
-        except Exception:
+        if domain_dir.name in _DOMAIN_BLOCKLIST or domain_dir.name.startswith("_"):
             continue
+        for module_path in domain_dir.rglob("*.py"):
+            if module_path.name == "__init__.py" or module_path.name.startswith("_"):
+                continue
+            rel_path = module_path.relative_to(_PACKAGE_ROOT)
+            module_name = ".".join(rel_path.with_suffix("").parts)
+            full_module_name = f"{_PACKAGE_NAME}.{module_name}"
 
-        for _name, _obj in vars(_module).items():
-            if (
-                callable(_obj)
-                and getattr(_obj, "_is_cobrabox_feature", False)
-                and getattr(_obj, "__module__", "") == full_module_name
-            ):
-                if _name in _discovered:  # pragma: no cover
-                    raise ValueError(
-                        f"Duplicate feature name '{_name}' while importing "
-                        f"module '{full_module_name}'."
-                    )
-                _discovered[_name] = _obj
+            try:
+                module = _importlib.import_module(full_module_name)
+            except Exception:
+                continue
 
+            for name, obj in vars(module).items():
+                if (
+                    callable(obj)
+                    and getattr(obj, "_is_cobrabox_feature", False)
+                    and getattr(obj, "__module__", "") == full_module_name
+                ):
+                    if name in discovered:  # pragma: no cover
+                        raise ValueError(
+                            f"Duplicate feature name '{name}' while importing "
+                            f"module '{full_module_name}'."
+                        )
+                    discovered[name] = obj
+    return discovered
+
+
+_discovered = _discover()
 globals().update(_discovered)
 __all__ = [*sorted(_discovered.keys())]  # noqa: PLE0604
