+++ src/cobrabox/spectral/__init__.py
@@ -1,0 +1,14 @@
+# Spectral features: what's happening in frequency space.
+from .band_power import BandPower
+from .continuous_wavelet_transform import ContinuousWaveletTransform
+from .cordance import Cordance
+from .discrete_wavelet_transform import DiscreteWaveletTransform
+from .spectrogram import Spectrogram
+
+__all__ = [
+    "BandPower",
+    "ContinuousWaveletTransform",
+    "Cordance",
+    "DiscreteWaveletTransform",
+    "Spectrogram",
+]
