+++ src/cobrabox/surrogates/__init__.py
@@ -1,0 +1,4 @@
+# Surrogate features: statistical null-hypothesis methods.
+from .fourier_transform_surrogates import FourierTransformSurrogates
+
+__all__ = ["FourierTransformSurrogates"]
