+++ src/cobrabox/windowing/__init__.py
@@ -1,1 +1,7 @@
-# Windowing features: windowing and aggregation
+# Windowing features: segmentation and aggregation over time.
+from .concat_aggregate import ConcatAggregate
+from .mean_aggregate import MeanAggregate
+from .sliding_window import SlidingWindow
+from .sliding_window_reduce import SlidingWindowReduce
+
+__all__ = ["ConcatAggregate", "MeanAggregate", "SlidingWindow", "SlidingWindowReduce"]
